requirements.txt
torch
torchvision
pandas
numpy
opencv-python
scikit-learn
matplotlib
tensorboard
streamlit
pyyaml
tqdm
